Debug `yy`, `dd`, and modularized code slightly
Three main things:
  1. Separated NormalEditor and InsertEditor into two different files with
Slate wrappers, but passed the same `value` prop from `App`, and `onChange`
also modifies `App`'s `value` state.
  2. Added a few more Normal mode commands like `p`, `P`, `d0`, `y0`, `d$`,
`y$`, `u`, `ctrl+r`.
  3. Debug `yy` and `dd`'s line-determination scheme.

There is one main issue with the implementation of `u` and `ctrl+r`, that
is, undo and redo. That is, `editor.undo()` and `editor.redo()` seem to be
affected by the structure of Slate nodes, though I have not checked this
yet. So the current behavior of deleting a line and using `editor.undo()`
results in some sort of undefined behavior.