Implement syncing combined editors
This implementation is mostly adapted from Ben Awad's
"mini-google-docs-clone". Added a new event emitter dependency, `mitt`, and
created wrapper editor classes for the synchronization logic. The main idea
of this sync-editing is to use ref to DOM elements to keep track of the
editor being changed and the remote one(s) and apply each operation of the
one being changed to all remote ones.

The way this is implemented now allows one to add arbitrarily many
`SyncingEditor`s which will all be synced on every operation across all
editors, regardless of which mode (Normal or Insert) that they are in.